JPA Optimation

+++ core/src/main/java/io/boodskap/iot/spi/storage/policy/wrapper/dao/AlexaDAOImpl.java
@@ -22,6 +22,7 @@ import io.boodskap.iot.StorageException;
 import io.boodskap.iot.dao.AlexaDAO;
 import io.boodskap.iot.dao.EntityIterator;
 import io.boodskap.iot.model.IAlexa;
+import io.boodskap.iot.spi.storage.policy.PolicyManager;
 
 public class AlexaDAOImpl implements AlexaDAO<IAlexa>{
 	
@@ -31,7 +32,9 @@ public class AlexaDAOImpl implements AlexaDAO<IAlexa>{
 		this.impl = impl;
 	}
 
+	@Override
 	public IAlexa create(String domainKey, String alexaId) {
+		PolicyManager.checkWriteAccess(domainKey);
 		return impl.create(domainKey, alexaId);
 	}
 
@@ -40,52 +43,75 @@ public class AlexaDAOImpl implements AlexaDAO<IAlexa>{
 		return impl.clazz();
 	}
 
+	@Override
 	public void createOrUpdate(IAlexa e) throws StorageException {
+		PolicyManager.checkWriteAccess(e.getDomainKey());
 		impl.createOrUpdate(e);
 	}
 
+	@Override
 	public EntityIterator<IAlexa> load() throws StorageException {
+		PolicyManager.checkAdminAccess();
 		return impl.load();
 	}
 
+	@Override
 	public EntityIterator<IAlexa> load(String domainKey) throws StorageException {
+		PolicyManager.checkReadAccess(domainKey);
 		return impl.load(domainKey);
 	}
 
+	@Override
 	public long count() throws StorageException {
+		PolicyManager.checkAdminAccess();
 		return impl.count();
 	}
 
+	@Override
 	public IAlexa get(String domainKey, String alexaId) throws StorageException {
+		PolicyManager.checkReadAccess(domainKey);
 		return impl.get(domainKey, alexaId);
 	}
 
+	@Override
 	public long count(String domainKey) throws StorageException {
+		PolicyManager.checkReadAccess(domainKey);
 		return impl.count(domainKey);
 	}
 
+	@Override
 	public void delete(String domainKey, String alexaId) throws StorageException {
+		PolicyManager.checkDeleteAccess(domainKey);
 		impl.delete(domainKey, alexaId);
 	}
 
+	@Override
 	public void delete(String domainKey) throws StorageException {
+		PolicyManager.checkDeleteAccess(domainKey);
 		impl.delete(domainKey);
 	}
 
+	@Override
 	public Collection<IAlexa> list(String domainKey, int page, int pageSize) throws StorageException {
+		PolicyManager.checkReadAccess(domainKey);
 		return impl.list(domainKey, page, pageSize);
 	}
 
-	public Collection<IAlexa> listNext(String domainKey, String alexaId, int page, int pageSize)
-			throws StorageException {
+	@Override
+	public Collection<IAlexa> listNext(String domainKey, String alexaId, int page, int pageSize) throws StorageException {
+		PolicyManager.checkReadAccess(domainKey);
 		return impl.listNext(domainKey, alexaId, page, pageSize);
 	}
 
+	@Override
 	public Collection<IAlexa> search(String domainKey, String query, int pageSize) throws StorageException {
+		PolicyManager.checkReadAccess(domainKey);
 		return impl.search(domainKey, query, pageSize);
 	}
 
+	@Override
 	public void delete() throws StorageException {
+		PolicyManager.checkAdminAccess();
 		impl.delete();
 	}
